Fix initialiser location
